+++ src/main/java/com/example/demo/OrderRepository.java
@@ -3,6 +3,6 @@ package com.example.demo;
 import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
 
 //wrapper around ReactiveCrudRepository
-public interface OrderRepository extends ReactiveMongoRepository<Order, Integer>{
+public interface OrderRepository extends ReactiveMongoRepository<Order, Long>{
 
 }
